Moved all of the record handling to its own module. Moved the stderr macro to the counter crate.
